Cascade/pdf_summarizer/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.24
langchain-openai>=0.0.5
python-dotenv==1.0.0
openai==1.3.5
pypdf==3.17.1
chromadb==0.4.18
fpdf==1.7.2
